requirements.txt
yfinance==0.2.54
numpy==1.26.4
pandas==2.2.2
scikit-learn==1.5.1
tensorflow==2.18.0
python-dateutil==2.9.0.post0
streamlit==1.37.1
matplotlib==3.9.2